move repeated code to common place
